chore: move particle options
