+++ src/app/_config/ParticleOptions.ts
